requirements.txt
alembic==1.4.2
atomicwrites==1.4.0
attrs==19.3.0
bleach==3.1.5
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-PageDown==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.2
pip-review==1.1.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-slugify==4.0.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.3.1
